:art: :white_check_mark: refactor tests into scripts | add tests for defaults
