+++ sdk_ooti_invoicing_tests.py
@@ -1048,525 +1048,6 @@ class Tests(unittest.TestCase):
 
         self.assertEqual(res['status'], 204)
 
-##### DELIVERABLES #####
-
-    #### Zones ####
-    def _create_zone_return_pk(self):
-        """ Create a zone and return the pk 
-
-        :return: {"pk": zone_pk, "area_pk": area_pk}
-        """
-
-        area_pk = self._create_area_return_pk()
-
-        data = {
-            "name": "UNITTEST",
-            "area": area_pk,
-            "progress": 0,
-            "is_annex": True,
-            "internal_code": "string",
-            "client_code": "string",
-            "surface_area": 0,
-            "default_surface_price": 0,
-            "num_units": 0
-        }
-
-        res = my_account.Deliverables.create_zone(area_pk, data)
-        return {"pk": res['data']['id'], "area_pk": area_pk}
-
-    def test_export_zones(self):
-        """ Test that 200 is returned """
-        res = my_account.Deliverables.export_phase(project_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    def test_get_zones_list(self):
-        """ Test that 200 is returned """
-        area_pk = self._create_area_return_pk()
-
-        res = my_account.Deliverables.get_zones_list(area_pk)
-        my_account.Deliverables.delete_area(area_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    def test_create_zone(self):
-        """ Test that 201 is returned """
-
-        area_pk = self._create_area_return_pk()
-
-        data = {
-            "name": "string",
-            "area": area_pk,
-            "progress": 0,
-            "is_annex": True,
-            "internal_code": "string",
-            "client_code": "string",
-            "surface_area": 0,
-            "default_surface_price": 0,
-            "num_units": 0
-        }
-
-        res = my_account.Deliverables.create_zone(area_pk, data)
-
-        self.assertEqual(res['status'], 201)
-
-    def test_get_zone_details(self):
-        """ Test that 200 is returned """
-        res_create = self._create_zone_return_pk()
-
-        res = my_account.Deliverables.get_zone_details(res_create['pk'])
-
-        my_account.Deliverables.delete_zone(res_create['pk'])
-        my_account.Deliverables.delete_area(res_create['area_pk'])
-
-        self.assertEqual(res['status'], 200)
-
-    def test_update_zone(self):
-        """ Test that 200 is returned """
-        res_create = self._create_zone_return_pk()
-
-        data_up = {
-            "name": "update"
-        }
-
-        res = my_account.Deliverables.update_zone(res_create['pk'], data_up)
-
-        my_account.Deliverables.delete_zone(res_create['pk'])
-        my_account.Deliverables.delete_area(res_create['area_pk'])
-
-        self.assertEqual(res['status'], 200)
-
-    def test_delete_zone(self):
-        """ Test that 204 is returned """
-        area_pk = self._create_area_return_pk()
-
-        data = {
-            "name": "string",
-            "area": area_pk,
-            "progress": 0,
-            "is_annex": True,
-            "internal_code": "string",
-            "client_code": "string",
-            "surface_area": 0,
-            "default_surface_price": 0,
-            "num_units": 0
-        }
-
-        zone_pk = my_account.Deliverables.create_zone(area_pk, data)['data']['id']
-        res = my_account.Deliverables.delete_zone(zone_pk)
-
-        my_account.Deliverables.delete_area(area_pk)
-
-        self.assertEqual(res['status'], 204)
-
-    #### Areas ####
-
-    def _create_area_return_pk(self):
-        """ Create an area and return the pk """
-
-        area = {
-            "name": "test",
-            "project": project_pk,
-            "surface_area": 30
-        }
-
-        return my_account.Deliverables.create_areas(project_pk, area)['data']['id']
-
-    def test_get_areas_list(self):
-        """ Test that 200 is returned """
-        # * OK
-        res = my_account.Deliverables.get_areas_list(project_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    def test_create_area(self):
-        """ Test that 201 is returned """
-        # * OK
-
-        area = {
-            "name": "UNITTEST",
-            "project": project_pk,
-            "surface_area": 30
-        }
-
-        res = my_account.Deliverables.create_areas(project_pk, area)
-        my_account.Deliverables.delete_area(res['data']['id'])
-
-        self.assertEqual(res['status'], 201)
-
-    def test_get_area_details(self):
-        """ Test that 200 is returned """
-        # * OK
-
-        area_pk = self._create_area_return_pk()
-
-        res = my_account.Deliverables.get_areas_details(area_pk)
-        my_account.Deliverables.delete_area(area_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    # def test_update_area(self):
-    #     """ Test that 200 is returned """
-    #     #! Do not pass (403)
-
-    #     area_pk = self._create_area_return_pk()
-
-    #     area_up = {
-    #         "name": "UPDATE",
-    #     }
-
-    #     res = my_account.Deliverables.create_areas(area_pk, area_up)
-    #     my_account.Deliverables.delete_area(area_pk)
-
-    #     self.assertEqual(res['status'], 200)
-
-    def test_delete_area(self):
-        """ Test that 204 is returned """
-        # * OK
-
-        area_pk = self._create_area_return_pk()
-        res = my_account.Deliverables.delete_area(area_pk)
-
-        self.assertEqual(res['status'], 204)
-
-    #### Phases ####
-    def test_get_phases_list(self):
-        """ Test that 200 is returned """
-        # * OK
-
-        res = my_account.Deliverables.get_phases_list(project_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    def test_get_phases_projections_list(self):
-        """ Test that 200 is returned """
-        # * OK
-
-        res = my_account.Deliverables.get_phases_projections_list(project_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    def test_export_phase(self):
-        """ Test that 200 is returned """
-        # * OK
-
-        res = my_account.Deliverables.export_phase(project_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    #### Milestone ####
-    def _create_milestone_return_pk(self):
-        """ Create a milestone and return the pk """
-
-        data = {
-            "title": "UNITTEST",
-            "project": project_pk,
-            "date": "30-04-2021",
-            "description": "string",
-            "in_timeline": True
-        }
-
-        return my_account.Deliverables.create_milestone(data)['data']['pk']
-
-    def test_get_milestones_list(self):
-        """ Test that 200 is returned """
-        # * OK
-
-        res = my_account.Deliverables.get_milestones_list()
-
-        self.assertEqual(res['status'], 200)
-
-    # def test_get_milestone_details(self):
-    #     """ Test that 200 is returned """
-    #     #! Do not pass, 403
-
-    #     milestone_pk = self._create_milestone_return_pk()
-    #     res = my_account.Deliverables.get_milestone_details(milestone_pk)
-
-    #     self.assertEqual(res['status'], 200)
-
-    def test_create_milestone(self):
-        """ Test that 201 is returned """
-        # * OK
-
-        data = {
-            "title": "UNITTEST",
-            "project": project_pk,
-            "date": "30-04-2021",
-            "description": "string",
-            "in_timeline": True
-        }
-
-        res = my_account.Deliverables.create_milestone(data)
-
-        self.assertEqual(res['status'], 201)
-
-    def test_update_milestone(self):
-        """ Test that 200 is returned """
-        # * OK
-
-        milestone_pk = self._create_milestone_return_pk()
-
-        data = {
-            "title": "UPDATE"
-        }
-
-        res = my_account.Deliverables.update_milestone(milestone_pk, data)
-
-        self.assertEqual(res['status'], 200)
-
-    #### Fees ####
-    def _create_fee_return_pk(self):
-        """ Create a fee and return pk """
-
-        data = {
-            "title": "string",
-            "amount_base": 0,
-            "amount_current": 0,
-            "progress": 0,
-            "in_timeline": True
-        }
-
-        return my_account.Deliverables.create_fee(project_pk, data)['data']['pk']
-
-    def test_get_fees_bracket_list(self):
-        """ Test that 200 is returned """
-        res = my_account.Deliverables.get_fees_bracket_list(project_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    def test_export_project_fees(self):
-        """ Test that 200 is returned """
-        res = my_account.Deliverables.export_project_fees(project_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    def test_get_fee_project_version_list(self):
-        """ Test that 200 is returned """
-        res = my_account.Deliverables.get_fee_project_version_list()
-
-        self.assertEqual(res['status'], 200)
-
-    def test_get_fees_list(self):
-        """ Test that 200 is returned """
-        res = my_account.Deliverables.get_fees_list(project_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    def test_create_fee(self):
-        """ Test that 201 is returned """
-
-        data = {
-            "title": "string",
-            "amount_base": 0,
-            "amount_current": 0,
-            "progress": 0,
-            "in_timeline": True
-        }
-
-        res = my_account.Deliverables.create_fee(project_pk, data)
-
-        self.assertEqual(res['status'], 201)
-
-    def test_get_fees_projection_list(self):
-        """ Test that 200 is returned """
-        res = my_account.Deliverables.get_fees_projection_list(project_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    def test_get_fees_project_list(self):
-        """ Test that 200 is returned """
-        res = my_account.Deliverables.get_fees_project_list()
-
-        self.assertEqual(res['status'], 200)
-
-    def test_validate_fees_costs(self):
-        """ Test that 200 is returned """
-        res = my_account.Deliverables.validation_fees_costs(project_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    def test_validate_fees_ffne(self):
-        """ Test that 200 is returned """
-        res = my_account.Deliverables.validation_fees_ffne(project_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    def test_validate_fees_production(self):
-        """ Test that 200 is returned """
-        res = my_account.Deliverables.validation_fees_production(project_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    def test_get_fees_zones_list(self):
-        """ Test that 200 is returend """
-        res = my_account.Deliverables.get_fees_zones_list()
-
-        self.assertEqual(res['status'], 200)
-
-    #### Plans ####
-    # def _create_plan_return_pk(self):
-    #     """ Create plan and return pk """
-
-    #     zone_pk = self._create_zone_return_pk()
-
-    #     data = {
-    #         "zone": zone_pk['pk'],
-    #         "name_fr": "string",
-    #         "name_en": "string",
-    #         "plan_format": "string",
-    #         "scale": "string",
-    #         "level": "string",
-    #         "lot": 0,
-    #         "is_default": True,
-    #         "progress": 0,
-    #         "sub_zone_code": "string",
-    #         "plan_code": "string",
-    #         "project": project_pk,
-    #         "area": zone_pk['area_pk'],
-    #         "code": "string",
-    #         "custom_field_1": "string",
-    #         "custom_field_2": "string",
-    #         "custom_field_3": "string"
-    #     }
-
-    #     res = my_account.Deliverables.create_plan(project_pk, data)['data']['id']
-
-    #     return {"pk": res, "area_pk": area_pk, "zone_pk": zone_pk}
-
-    def test_get_plans_list_action(self):
-        """ Test that 200 is returned """
-        res = my_account.Deliverables.get_plans_list_action(project_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    def test_create_plans_list_action(self):
-        """ Test that 201 is returned """
-        res = my_account.Deliverables.create_plans_list_action(project_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    def test_get_plans_list(self):
-        """ Test that 200 is returned """
-        res = my_account.Deliverables.get_plans_list(project_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    # def test_create_plan(self):
-    #     """ Test that 201 is returned """
-
-    #     zone_pk = self._create_zone_return_pk()
-
-    #     data = {
-    #         "zone": zone_pk['pk'],
-    #         "name_fr": "string",
-    #         "name_en": "string",
-    #         "plan_format": "string",
-    #         "scale": "string",
-    #         "level": "string",
-    #         "lot": 0,
-    #         "is_default": True,
-    #         "progress": 0,
-    #         "sub_zone_code": "string",
-    #         "plan_code": "string",
-    #         "project": project_pk,
-    #         "area": zone_pk['area_pk'],
-    #         "code": "client"
-    #     }
-
-    #     res = my_account.Deliverables.create_plan(project_pk, data)
-
-    #     my_account.Deliverables.delete_zone(zone_pk['pk'])
-    #     my_account.Deliverables.delete_area(zone_pk['area_pk'])
-    #     # my_account.Deliverables.delete_plan(res['data']['pk'])
-
-    #     self.assertEqual(res['status'], 201)
-
-    #### Prescription ####
-    def _create_prescription_return_pk(self):
-        """ Create a prescription and return pk 
-
-        :return: {"pk": pk, "area_pk": area_pk, "fee_pk": fee_pk}
-
-        """
-
-        fee_pk = self._create_fee_return_pk()
-        area_pk = self._create_area_return_pk()
-
-        data = {
-            "fee_pct": "10",
-            "fee": fee_pk,
-            "date": "04-05-2021",
-            "description": "UNITTEST",
-            "area": area_pk,
-        }
-
-        pk = my_account.Deliverables.create_prescription(project_pk, data)['data']['id']
-        return {"pk": pk, "area_pk": area_pk, "fee_pk": fee_pk}
-
-    def test_get_prescription_list(self):
-        """ Test that 200 is returned """
-        res = my_account.Deliverables.get_prescriptions_list(project_pk)
-
-        self.assertEqual(res['status'], 200)
-
-    def test_create_prescription(self):
-        """ Test that 201 is returned """
-
-        fee_pk = self._create_fee_return_pk()
-        area_pk = self._create_area_return_pk()
-
-        data = {
-            "fee_pct": "10",
-            "fee": fee_pk,
-            "date": "04-05-2021",
-            "description": "UNITTEST",
-            "area": area_pk,
-        }
-
-        res = my_account.Deliverables.create_prescription(project_pk, data)
-        my_account.Deliverables.delete_prescription(res['data']['id'])
-        my_account.Deliverables.delete_area(area_pk)
-
-        self.assertEqual(res['status'], 201)
-
-    def test_get_prescription_details(self):
-        """ Test that 200 is returned """
-
-        res_creation = self._create_prescription_return_pk()
-
-        res = my_account.Deliverables.get_prescriptions_details(res_creation['pk'])
-        my_account.Deliverables.delete_prescription(res_creation['pk'])
-        my_account.Deliverables.delete_area(res_creation['area_pk'])
-
-        self.assertEqual(res['status'], 200)
-
-    def test_update_prescription(self):
-        """ Test that 200 is returned """
-
-        res_creation = self._create_prescription_return_pk()
-
-        data = {
-            "description": "UPDATED"
-        }
-
-        res = my_account.Deliverables.update_prescriptions(res_creation['pk'], data)
-        my_account.Deliverables.delete_prescription(res_creation['pk'])
-        my_account.Deliverables.delete_area(res_creation['area_pk'])
-
-        self.assertEqual(res['status'], 200)
-
-    def test_delete_prescription(self):
-        """ Test that 204 is returned """
-
-        res_creation = self._create_prescription_return_pk()
-
-        res = my_account.Deliverables.delete_prescription(res_creation['pk'])
-
-        my_account.Deliverables.delete_area(res_creation['area_pk'])
-
-        self.assertEqual(res['status'], 204)
-
 
 if __name__ == '__main__':
     unittest.main()
